+++ src/features/invoices/components/InvoiceFilters.jsx
@@ -1,0 +1,130 @@
+import React from 'react';
+
+const InvoiceFilters = ({ 
+  filters, 
+  setFilters, 
+  searchInput, 
+  setSearchInput, 
+  sortBy, 
+  setSortBy, 
+  sortDir, 
+  setSortDir, 
+  pageSize, 
+  setPageSize, 
+  setPage 
+}) => {
+  const handleClearAll = () => {
+    setFilters({ status: 'all', dateFrom: '', dateTo: '', search: '' });
+    setSearchInput('');
+    setPage(1);
+  };
+
+  return (
+    <div className="bg-white rounded-xl shadow-sm border border-gray-200 p-6">
+      <div className="flex items-center justify-between mb-6">
+        <h3 className="text-lg font-semibold text-gray-900">Filters & Search</h3>
+        <button onClick={handleClearAll} className="text-sm text-gray-500 hover:text-gray-700 underline">
+          Clear All
+        </button>
+      </div>
+      
+      <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-4">
+        <div>
+          <label className="block text-sm font-medium text-gray-700 mb-2">Status</label>
+          <select
+            value={filters.status}
+            onChange={(e) => setFilters(prev => ({ ...prev, status: e.target.value }))}
+            className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+          >
+            <option value="all">All Status</option>
+            <option value="draft">Draft</option>
+            <option value="pending">Pending</option>
+            <option value="partial">Partial</option>
+            <option value="paid">Paid</option>
+            <option value="overdue">Overdue</option>
+            <option value="cancelled">Cancelled</option>
+          </select>
+        </div>
+        
+        <div>
+          <label className="block text-sm font-medium text-gray-700 mb-2">From Date</label>
+          <input
+            type="date"
+            value={filters.dateFrom}
+            onChange={(e) => setFilters(prev => ({ ...prev, dateFrom: e.target.value }))}
+            className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+          />
+        </div>
+        
+        <div>
+          <label className="block text-sm font-medium text-gray-700 mb-2">To Date</label>
+          <input
+            type="date"
+            value={filters.dateTo}
+            onChange={(e) => setFilters(prev => ({ ...prev, dateTo: e.target.value }))}
+            className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+          />
+        </div>
+        
+        <div className="sm:col-span-2 lg:col-span-1">
+          <label className="block text-sm font-medium text-gray-700 mb-2">Search</label>
+          <div className="relative">
+            <input
+              type="text"
+              value={searchInput}
+              onChange={(e) => { setPage(1); setSearchInput(e.target.value); }}
+              placeholder="Invoice number or customer..."
+              className="w-full pl-10 pr-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+            />
+            <svg className="absolute left-3 top-2.5 h-5 w-5 text-gray-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z" />
+            </svg>
+          </div>
+        </div>
+      </div>
+      
+      <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4 mt-4">
+        <div>
+          <label className="block text-sm font-medium text-gray-700 mb-2">Sort By</label>
+          <select
+            value={sortBy}
+            onChange={(e) => { setPage(1); setSortBy(e.target.value); }}
+            className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+          >
+            <option value="date">Date</option>
+            <option value="amount">Amount</option>
+            <option value="customer">Customer</option>
+            <option value="status">Status</option>
+          </select>
+        </div>
+        
+        <div>
+          <label className="block text-sm font-medium text-gray-700 mb-2">Order</label>
+          <select
+            value={sortDir}
+            onChange={(e) => { setPage(1); setSortDir(e.target.value); }}
+            className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+          >
+            <option value="desc">Newest First</option>
+            <option value="asc">Oldest First</option>
+          </select>
+        </div>
+        
+        <div>
+          <label className="block text-sm font-medium text-gray-700 mb-2">Items Per Page</label>
+          <select
+            value={pageSize}
+            onChange={(e) => { setPage(1); setPageSize(Number(e.target.value)); }}
+            className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+          >
+            <option value={10}>10 items</option>
+            <option value={20}>20 items</option>
+            <option value={50}>50 items</option>
+          </select>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default InvoiceFilters;
