+++ src/features/invoices/index.js
@@ -1,0 +1,5 @@
+// Invoice feature exports
+export { default as InvoiceSummaryCards } from './components/InvoiceSummaryCards';
+export { default as InvoiceFilters } from './components/InvoiceFilters';
+export { default as InvoiceTable } from './components/InvoiceTable';
+export { useInvoices } from './hooks/useInvoices';
