+++ src/features/pos/index.js
@@ -1,0 +1,4 @@
+// POS feature exports
+export { default as POSItemList } from './components/POSItemList';
+export { default as POSBillingSummary } from './components/POSBillingSummary';
+export { usePOSCalculations } from './hooks/usePOSCalculations';
