+++ src/features/shared/index.js
@@ -1,0 +1,4 @@
+// Shared components exports
+export { default as LoadingSpinner } from './components/LoadingSpinner';
+export { default as EmptyState } from './components/EmptyState';
+export { default as ErrorState } from './components/ErrorState';
